chore(apply-patch) move invocation parsing
lib.rs has grown quite large, and mixes two responsibilities:
1. executing patch operations
2. parsing apply_patch invocations via a shell command

This PR splits out (2) into its own file, so we can work with it more easily. We are explicitly NOT moving tests in this PR, to ensure behavior stays the same and we can avoid losing coverage via merge conflicts. Tests are moved in a subsequent PR.
